Move Auth.tsx to system folder
