+++ src-mobile-app/src/app/(auth)/_layout.tsx
@@ -1,5 +1,5 @@
 import { Redirect, Slot } from "expo-router";
-import { useAuth } from "../../lib/auth";
+import { useAuth } from "../../features/system/Auth";
 
 export default function AuthLayout() {
   const { session, isReady } = useAuth();
