+++ src-mobile-app/src/features/system/Auth.tsx
@@ -2,9 +2,9 @@ import 'react-native-url-polyfill/auto';
 import { Session, User } from '@supabase/supabase-js';
 import { createContext, useContext, useState } from 'react';
 import { Alert } from 'react-native';
-import { useTypedDispatch } from '../store/store';
-import { setUserID } from '../store/systemSlice';
-import { supabase } from './supabase';
+import { useTypedDispatch } from '../../store/store';
+import { setUserID } from '../../store/systemSlice';
+import { supabase } from '../../lib/supabase';
 
 export interface AuthSessionState {
   session: Session | null
